메인/상세 예외처리
