+++ src/main/java/com/group15/CreamCloneBackend/domain/product/dto/MainReponseDto.java
@@ -2,6 +2,7 @@ package com.group15.CreamCloneBackend.domain.product.dto;
 
 import lombok.AllArgsConstructor;
 import lombok.Getter;
+import lombok.NoArgsConstructor;
 import lombok.Setter;
 
 import java.util.List;
@@ -9,6 +10,7 @@ import java.util.List;
 @Setter
 @Getter
 @AllArgsConstructor
+@NoArgsConstructor
 public class MainReponseDto {
     private List<MainDto> productList;
     private Long statusCode;
